Embedded DB example beautifications.
