+++ demo/EmbeddedDB/Makefile
