--- demo/EmbeddedStorage/demo.cc
@@ -1,89 +1,0 @@
-/*******************************************************************************
-The MIT License (MIT)
-
-Copyright (c) 2015 Dmitry "Dima" Korolev <dmitry.korolev@gmail.com>
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
-*******************************************************************************/
-
-#include "data_dictionary.h"
-
-#include "../../Sherlock/sherlock.h"
-#include "../../Bricks/dflags/dflags.h"
-
-using namespace current;
-
-DEFINE_int32(current_demo_port, 8889, "Local port to spawn the server on.");
-DEFINE_string(current_db_dir, ".current", "Local path to the data storage location.");
-DEFINE_string(current_db_filename, "data.json", "File name for the persisted data.");
-
-int main(int argc, char** argv) {
-  ParseDFlags(&argc, &argv);
-
-  const std::string persistence_file_name =
-      bricks::FileSystem::JoinPath(FLAGS_current_db_dir, FLAGS_current_db_filename);
-
-  auto stream = sherlock::Stream<Event, blocks::persistence::NewAppendToFile>("Data", persistence_file_name);
-
-  HTTP(FLAGS_current_demo_port).Register("/healthz", [](Request r) { r("OK\n"); });
-
-  // Schema.
-  using current::reflection::SchemaInfo;
-  using current::reflection::StructSchema;
-  using current::reflection::Language;
-
-  HTTP(FLAGS_current_demo_port).Register("/schema.cc", [](Request r) {
-    StructSchema schema;
-    schema.AddType<Event>();
-    r(schema.Describe(Language::CPP()), HTTPResponseCode.OK, "text/plain; charset=us-ascii");
-  });
-
-  HTTP(FLAGS_current_demo_port).Register("/schema.fs", [](Request r) {
-    StructSchema schema;
-    schema.AddType<Event>();
-    r(schema.Describe(Language::FSharp()), HTTPResponseCode.OK, "text/plain; charset=us-ascii");
-  });
-
-  HTTP(FLAGS_current_demo_port).Register("/schema.json", [](Request r) {
-    StructSchema schema;
-    schema.AddType<Event>();
-    r(schema.GetSchemaInfo());
-  });
-
-  // Subscribe.
-  HTTP(FLAGS_current_demo_port).Register("/data", stream);
-
-  // Publish.
-  HTTP(FLAGS_current_demo_port).Register("/publish", [&stream](Request r) {
-    if (r.method == "POST") {
-      try {
-        auto event = ParseJSON<Event>(r.body);
-        stream.Publish(std::move(event));
-        r("", HTTPResponseCode.NoContent);
-      } catch (const Exception& e) {
-        r(Error(e.What()), HTTPResponseCode.BadRequest);
-      }
-    } else {
-      r(Error("This request should be a POST."), HTTPResponseCode.MethodNotAllowed);
-    }
-  });
-
-  // Run forever.
-  HTTP(FLAGS_current_demo_port).Join();
-}
